python_api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymupdf>=1.24.2
google-generativeai>=0.7.2
python-dotenv>=1.0.1
python-multipart>=0.0.9
apscheduler>=3.11.2
